# Optimize tokenization performance with StringScanner (73% improvement)