--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -46,8 +46,8 @@ default_install_hook_types:
 # Default execution stage
 default_stages: [pre-commit]
 
-# Terminate checks on first failure
-fail_fast: true
+# Run all hooks even if some fail (allows multiple auto-fixes in one run)
+fail_fast: false
 
 # Ignore generated and dependency directories
 exclude: '^$'