--- a/lib/rhales/version.rb
+++ b/lib/rhales/version.rb
@@ -2,5 +2,5 @@
 
 module Rhales
   # Version information for the RSFC gem
-  VERSION = '0.2.5'
+  VERSION = '0.3.0'
 end